Renames and tweaks for consistent self-documentation and usage.

+++ bin/inc/vars.sh
@@ -2,38 +2,59 @@
 
 # Various utility routines. This is intended to be included in other scripts.
 
-DFLT_WDIR="$(cd "$(dirname "$(dirname "$(dirname "$0")")")"; pwd)"
-export WDIR="${WDIR-${DFLT_WDIR}}"]
-export BUILD="${BUILD-"${WDIR}/build"}"
-export BUILDTMP="${BUILD}/tmp"
-export DATA="${WDIR}/data"
-export SAVED="${WDIR}/saved"
+CMD="$0"
+export PI_INCLUDES="${PI_INCLUDES:-"$(cd "$(dirname "$0")"; pwd)"}"
+export PI_CMDS="${PI_CMDS:-"$(dirname "${PI_INCLUDES}")"}"
+DFLT_WDIR="$(dirname "${PI_CMDS}")"
+export PI_WORKDIR="${PI_WORKDIR:-"${DFLT_WDIR}"}"
+export PI_BUILD="${PI_BUILD:-"${PI_WORKDIR}/build"}"
+export PI_TMP="${PI_TMP:-"${PI_BUILD}/tmp"}"
+export PI_DATA="${PI_DATA:-"${PI_WORKDIR}/data"}"
+export PI_SAVED="${PI_SAVED:-"${PI_WORKDIR}/saved"}"
 
-export ROOTDIR="${BUILD}/root"
-export BOOTDIR="${ROOTDIR}/boot"
+export PI_ROOT="${PI_ROOT:-"${PI_BUILD}/root"}"
+export PI_BOOT="${PI_BOOT:-"${PI_ROOT}/boot"}"
 
 SELF=$$
 
-export VERBOSE="${VERBOSE}"
-export DEBUG="${DEBUG}"
+export PI_VERBOSE="${PI_VERBOSE}"
+export PI_VERBOSE="${PI_DEBUG}"
+
+# Show usage for the current or
+ usage() {
+     local cmd="${1:-"${CMD}"}"
+     local script="$(grep -E '^#### |^####$' "${cmd}" | sed -E -e 's/^#### ?/echo "/' -e 's/$/";/')"
+     if [ -z "${script}" ]; then
+        msg "The script ${cmd} lacks documentation."
+        msg "  Subcommand documentation is a set of comments beginning with '#### '."
+        msg "  These are stripped of the '#### ', and shell substitutions are performed,"
+        msg "  so help text can reference environment variables, etc."
+        msg "  Particlarly useful is the PI_INVOKER environment variable, which holds"
+        msg "  help for the words prior to the subcommand on the command line."
+        exit 0
+    else
+        eval "${script}" 1>&2
+        exit 0
+    fi
+ }
 
 msg() {
     echo "$@" 1>&2
 }
 
 verbose() {
-    test ! -z "${VERBOSE}${DEBUG}" && msg "$@"
+    test ! -z "${PI_VERBOSE}${PI_DEBUG}" && msg "$@"
     return 0
 }
 
 debug() {
-    test ! -z "${DEBUG}" && msg "$@"
+    test ! -z "${PI_DEBUG}" && msg "$@"
     return 0
 }
 
 # Unmount the specified directory
 do_unmount() {
-    local dir="$1"
+    local dir="${1:?}"
     if [ -e "${dir}" ] &&  grep -q "${dir}" /etc/mtab; then
         sync
 	    umount -l -d "${dir}"
@@ -43,49 +64,54 @@ do_unmount() {
 
 # Delete the loopback devices
 do_delete_loop() {
-    if [ -e "${LOOPDEV}" ]; then
-        kpartx -d "${LOOPDEV}"
-    fi
-    if [ -e "${LOOPDEV}" ]; then
-        # May already have been detached.
-        losetup --detach "${LOOPDEV}" 2>/dev/null
+    if [ ! -z "${PI_LOOPDEV}" ]; then
+        if [ -e "${PI_LOOPDEV}" ]; then
+            kpartx -d "${PI_LOOPDEV}"
+        fi
+        if [ -e "${PI_LOOPDEV}" ]; then
+            # May already have been detached.
+            losetup --detach "${PI_LOOPDEV}" 2>/dev/null
+        fi
+        unset PI_LOOPDEV
+        unset PI_BOOTDEV
+        unset PI_ROOTDEV
     fi
 }
 
 # Delete our temporary files
 do_delete_tmp() {
-    if [ -d "${BUILDTMP}" ]; then
-        debug "Removing previous temp ${BUILDTMP}"
-        rm -rf "${BUILDTMP}" 2>/dev/null
+    if [ -d "${PI_TMP:?}" ]; then
+        debug "Removing previous temp ${PI_TMP:?}"
+        rm -rf "${PI_TMP:?}" 2>/dev/null
     fi
 }
 
-# Unmount all our directories,m and clean up our temporary directory.
+# Unmount all our directories, and clean up our temporary directory.
 do_unmount_all() {
     do_delete_tmp
-    do_unmount "${BOOTDIR}"
-    do_unmount "${ROOTDIR}"
+    do_unmount "${PI_BOOT:?}"
+    do_unmount "${PI_ROOT:?}"
     sleep 1
     do_delete_loop
 }
 
-# Set LOOPDEV, ROOTDEV, and BOOTDEV to the raw devices
+# Set PI_LOOPDEV, PI_ROOTDEV, and PI_BOOTDEV to the raw devices
 # for the whole disk and its root and boot partitions.
 find_partitions() {
-    mapfile -t PARTS < <(kpartx -avs "${IMG}" | cut -d' ' -f3)
-    export BOOTDEV="/dev/mapper/${PARTS[0]}"
-    export ROOTDEV="/dev/mapper/${PARTS[1]}"
-    export LOOPDEV="/dev/$(echo "${PARTS[0]}" | sed -E 's/p[0-9]+$//')"
-    verbose Found LOOPDEV="${LOOPDEV}"
-    verbose Found BOOTDEV="${BOOTDEV}" BOOTDIR="${BOOTDIR}"
-    verbose Found ROOTDEV="${ROOTDEV}" ROOTDIR="${ROOTDIR}"
+    mapfile -t PARTS < <(kpartx -avs "${PI_IMAGE_FILE:?}" | cut -d' ' -f3)
+    export PI_BOOTDEV="/dev/mapper/${PARTS[0]}"
+    export PI_ROOTDEV="/dev/mapper/${PARTS[1]}"
+    export PI_LOOPDEV="/dev/$(echo "${PARTS[0]}" | sed -E 's/p[0-9]+$//')"
+    verbose Found PI_LOOPDEV="${PI_LOOPDEV}"
+    verbose Found PI_BOOTDEV="${PI_BOOTDEV}" PI_BOOT="${PI_BOOT}"
+    verbose Found PI_ROOTDEV="${PI_ROOTDEV}" PI_ROOT="${PI_ROOT}"
 }
 
 # Mount the specified directory on the specified location
 # Arrange to clean up if the mount fails.
 do_mount() {
-    local mapped="$1"
-    local mountpoint="$2"
+    local mapped="${1:?}"
+    local mountpoint="${2:?}"
     (
         mount -o loop "${mapped}" "${mountpoint}" \
         && verbose "${mapped} mounted on ${mountpoint}"
@@ -97,31 +123,31 @@ do_mount() {
 
 # Mount our partitions and our temporary work area.
 do_mount_all() {
-    do_unmount_all 2>/dev/null
-    mkdir -p "${BUILDTMP}" 2>/dev/null
-    verbose "Mounting partitions from ${IMG}"
+    do_unmount_all
+    mkdir -p "${PI_TMP:?}" 2>/dev/null
+    verbose "Mounting partitions from ${PI_IMAGE_FILE:?}"
     find_partitions
     trap "do_unmount_all" EXIT
-    do_mount "${ROOTDEV}" "${ROOTDIR}"
-    do_mount "${BOOTDEV}" "${BOOTDIR}"
+    do_mount "${PI_ROOTDEV:?}" "${PI_ROOT:?}"
+    do_mount "${PI_BOOTDEV:?}" "${PI_BOOT:?}"
 }
 
 # mkktmp dir
 # make a directory in our temporary workspace.
 # Takes the full relative or absolute path
 mktmp() {
-    rm -rf "$1" 2>/dev/null
-    mkdir -p "${1}"
+    rm -rf "${1:?}" 2>/dev/null
+    mkdir -p "${1:?}"
 }
 
 # copy file
 # copies from data to buildtmp
 copy() {
-    cp "${DATA}/$1" "${BUILDTMP}/$1"
+    cp "${PI_DATA}/${1:?}" "${PI_TMP}/$1"
 }
 
 copyUntilInternal() {
-    local terminator="${1-## Added}"
+    local terminator="${1:-## Added}"
      while read line; do
         if [ "$line" = "${terminator}" ]; then
             return 0
@@ -136,41 +162,64 @@ copyUntilInternal() {
 # but if the script is run again, the data is replaced,
 # not endlessly added.
 copyUntil() {
-    copyUntilInternal "${2-## Added}" <"${ROOTDIR}/$1" >"${BUILDTMP}/$1"
+    copyUntilInternal "${2-## Added}" <"${PI_ROOT:?}/$1" >"${PI_TMP:?}/$1"
 }
 
 # append file
 # Appends file from data to the destination (result in buildtmp)
 append() {
-    cp "${BUILDROOT}/$1" "${BUILDTMP}/$1"
+    cp "${BUILDROOT:?}/${1:?}" "${PI_TMP:?}/${1:?}"
     appendLine "$1" '## Added'
-    cat "${DATA}/$1" >>"${BUILDTMP}/$1"
+    cat "${PI_DATA:?}/${1:?}" >>"${PI_TMP:?}/${1:?}"
 }
 
 # appendLine file line
 # Result in buildtmp
 appendLine() {
-    echo "$2" >>"${BUILDTMP}/$1"
+    echo "${2:?}" >>"${PI_TMP:?}/${1:?}"
 }
 
 # install file [ perms ]
 # Install data from buildtmp area to the real filesystem
 install() {
-    local file="$1"
-    local perms="${2-644}"
+    local file="${1:?}"
+    local perms="${2:-644}"
     verbose "Installing ${file}"
-    cp "${ROOTDIR}/${file}" "${SAVED}/${file}"
-    cp "${BUILDTMP}/${file}" "${ROOTDIR}/${file}"
-    chmod "${perms}" "${ROOTDIR}/${file}"
+    cp "${PI_ROOT:?}/${file}" "${PI_SAVED:?}/${file}"
+    cp "${PI_TMP:?}/${file}" "${PI_ROOT:?}/${file}"
+    chmod "${perms}" "${PI_ROOT:?}/${file}"
 }
 
 # Install a user's home directory from pre-supplied data, taking care
 # to set the permissions on .ssh/ correctly.
 installHome() {
-    local user="$1"
+    local user="${1:?}"
     verbose "Installing user files for ${user}"
-    cp -a "${DATA}/home/${user}" "${ROOTDIR}/home/${user}"
-    cp -a "${DATA}/home/${user}/.ssh" "${ROOTDIR}/home/${user}/.ssh"
-    find "${ROOTDIR}/home/${user}/.ssh" -type d -exec chmod 700 {} \;
-    find "${ROOTDIR}/home/${user}/.ssh" -type f -exec chmod 600 {} \;
+    cp -a "${PI_DATA:?}/home/${user}" "${PI_ROOT:?}/home/${user}"
+    cp -a "${PI_DATA:?}/home/${user}/.ssh" "${PI_ROOT:?}/home/${user}/.ssh"
+    find "${PI_ROOT:?}/home/${user}/.ssh" -type d -exec chmod 700 {} \;
+    find "${PI_ROOT:?}/home/${user}/.ssh" -type f -exec chmod 600 {} \;
 }
+
+
+unset options_done
+while [ ! -z "$*" -a -z "$options_done" ]; do
+    case "$1" in
+        --help|'-?')
+            usage
+            exit 0
+            ;;
+        --debug|-d|-vv)
+            export PI_VERBOSE=yes
+            export PI_DEBUG=yes
+            shift
+            ;;
+        --verbose|-v)
+            export PI_VERBOSE=yes
+            shift
+            ;;
+        *)
+            options_done=true
+            ;;
+    esac
+done
